refactor: Storage trait
