+++ core/rpc/core/src/tests/sqlite/mod.rs
@@ -4,7 +4,9 @@ use common::{Context, Result};
 use sql::*;
 
 use ckb_jsonrpc_types::BlockView as JsonBlockView;
-use core_storage::{RelationalStorage, Storage};
+
+use core_storage::RelationalStorage;
+use protocol::storage::Storage;
 use xsql::rbatis::executor::RBatisTxExecutor;
 
 pub async fn delete_all_data(tx: &mut RBatisTxExecutor<'_>) -> Result<()> {
