+++ protocol/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod db;
 pub mod extension;
+pub mod storage;
